[FEAT] 영화 상세보기 수정

+++ src/main/java/com/insidemovie/backend/api/movie/dto/TmdbMovieDto.java
@@ -32,7 +32,7 @@ public class TmdbMovieDto {
     private LocalDate releaseDate; //개봉일
 
     @JsonProperty("genre_ids")
-    private List<Integer> genreIds; //장르
+    private List<Long> genreIds; //장르
 
     @JsonProperty("original_language")
     private String originalLanguage; //국가
